+++ Magic/src/main/java/com/elmakers/mine/bukkit/crafting/MagicCampfireRecipe.java
@@ -6,13 +6,13 @@ import org.bukkit.inventory.Recipe;
 import com.elmakers.mine.bukkit.magic.MagicController;
 import com.elmakers.mine.bukkit.utility.CompatibilityLib;
 
-public class MagicFurnaceRecipe extends MagicCookingRecipe {
-    protected MagicFurnaceRecipe(String key, MagicController controller) {
+public class MagicCampfireRecipe extends MagicCookingRecipe {
+    protected MagicCampfireRecipe(String key, MagicController controller) {
         super(key, controller);
     }
 
     @Override
     protected Recipe createRecipe(ItemStack item) {
-        return CompatibilityLib.getCompatibilityUtils().createFurnaceRecipe(key, item, ingredient.getItemStack(1), ignoreDamage, experience, cookingTime);
+        return CompatibilityLib.getCompatibilityUtils().createCampfireRecipe(key, item, ingredient.getItemStack(1), ignoreDamage, experience, cookingTime);
     }
 }