+++ Magic/CHANGELOG.md
@@ -5,7 +5,7 @@
 ## Additions
 
  - New external example added, `/mconfig fetch httyd` - a "How to Train Your Dragon" config pack from Night!
- - Add `type: furnace` option for recipes, 1.14+ only
+ - Add `type:` option for recipes, 1.14+ only, can be any of: `furnace`,`campfire`,`blasting`,`smoking`
  - Add `EntityAnimation` action, can make players or mobs swing their arms
  - New EffectLib parameters to support 1.17 particles:
    - `to_color`: Used for the dust_color_transition particlea